Real-execution verification: fix broken nav routes, Carbon TTL bug, apply ADR-0013 guard
This pass actually ran InfoDot against real PHP 8.5 and PostgreSQL for the
first time — composer install ran clean, and php artisan migrate ran clean
on the first try with no ordering bugs. php artisan test surfaced several
real, previously-uncaught bugs:

1. tests/TestCase.php never called withoutVite(), unlike every sibling
   platform (Dot.Sheet, Dot.docs) — every view-rendering test failed on a
   missing Vite manifest that's never built in this environment. Added
   withoutVite() to setUp(), matching the established sibling convention.

2. EcosystemTokenController::issue(), EcosystemWidget::launch(), and
   DotSwitcher::launch() all passed config('ecosystem.handoff_ttl', 5)
   straight into Carbon's addMinutes() without casting — env() returns
   numeric .env values as strings, and Carbon 3's rawAddUnit() requires
   int|float, so every ecosystem SSO handoff token issuance threw a
   TypeError. Fixed by casting to (int) at all three call sites.

3. layouts/app.blade.php's sidebar nav referenced route('dashboard') and
   route('ecosystem.widget') — neither name was ever defined anywhere in
   routes/web.php, in this repo's entire git history. This meant every
   authenticated page render threw RouteNotFoundException in real usage,
   not just tests. Fixed by naming the existing /home route 'dashboard'
   (it already serves as the authenticated landing page via
   HomeController@index) and adding a new /platforms route wired to the
   existing App\Livewire\EcosystemWidget component, named 'ecosystem.widget'.

4. The same layout also called route('profile.show') with no parameter,
   but that route requires a {id} segment — fixed both call sites to pass
   auth()->id().

5. <x-banner /> in the same layout used the wrong component name: this
   app's AppServiceProvider registers all Jetstream Blade components under
   a jet-* prefix, so the actual banner component is jet-banner, not
   banner. Fixed to <x-jet-banner />.

After these fixes, php artisan test passed all 93 tests cleanly (0
failures; most are flagged "deprecated" only because of a PHP 8.5
deprecation notice on the stock Laravel PDO::MYSQL_ATTR_SSL_CA constant in
config/database.php, unrelated to any of the above and left untouched as
unmodified framework boilerplate).

Also applied the Dot.Brain adr/ADR-0013 idempotent-migration guard to this
platform's six shared Jetstream-core migrations (users, two-factor
columns, personal_access_tokens, teams, team_user, team_invitations),
wrapping each Schema::create/addColumn call in a Schema::hasTable/hasColumn
check so this platform's copies are safe to co-execute against the same
shared infodot database as any other Dot platform, regardless of run
order. Re-verified against a fresh empty database after guarding: same
clean migrate and the same 93-test result, confirming no behavior change.

InfoDot has no wiki.md (confirmed absent, unlike its siblings), so this
pass has no wiki changelog entry to add.

+++ app/Http/Controllers/Api/EcosystemTokenController.php
@@ -13,7 +13,7 @@ class EcosystemTokenController extends Controller
     {
         $user = $request->user();
 
-        $ttlMinutes = config('ecosystem.handoff_ttl', 5);
+        $ttlMinutes = (int) config('ecosystem.handoff_ttl', 5);
 
         $token = $user->createToken(
             'ecosystem-handoff',
